server/balances_history_cacher/requirements.txt
aiohttp==3.8.1
apscheduler==3.8.0
django-cors-headers==3.9.0
django-encrypted-model-fields==0.6.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
pip-chill==1.0.1
requests==2.26.0
pymongo==3.12.1